requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.9.0
anthropic>=0.40.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
